+++ PCMC_lib/include/Energy/Tanh.h
@@ -33,7 +33,7 @@ public:
     double energy(double distance) const;
     double energyAllSites(const Polymer& polymer) const;
     double energyIfSiteChanged(int site, int size, const Vector* r) const;
-    virtual double energyIfSiteChangedCL(int site, int size, const double* r) const;
+    virtual double energyIfSiteChangedCL(int site, int size, const float* r) const;
     virtual void initCL() const;
     virtual void cleanCL() const;
